+++ REST PRO/0b407b41-9285-47de-902f-94a453bbde4a-10a3c895-1df1-43da-8e32-483ab136689c-Vehicle_Repair_Hub-Pranesh-Ramachandran/springapp/src/main/java/com/examly/springapp/model/UserController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.data.domain.Page;
 
 import java.util.List;
-
 @RestController
 @RequestMapping("/users")
 public class UserController {
@@ -18,61 +17,54 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-   
     @PostMapping
     public ResponseEntity<User> createUser(@RequestBody User user) {
         User savedUser = userService.createUser(user);
         return new ResponseEntity<>(savedUser, HttpStatus.CREATED);
     }
 
-    
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable int id) {
+    public ResponseEntity<User> getUserById(@PathVariable Long id) {
         User user = userService.getUserById(id);
-        return user != null ? 
-                new ResponseEntity<>(user, HttpStatus.OK) : 
+        return user != null ?
+                new ResponseEntity<>(user, HttpStatus.OK) :
                 new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    
     @GetMapping
     public ResponseEntity<List<User>> getAllUsers() {
         List<User> userList = userService.getAllUsers();
         return new ResponseEntity<>(userList, HttpStatus.OK);
     }
 
-    
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable int id, @RequestBody User updatedUser) {
+    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User updatedUser) {
         User updated = userService.updateUser(id, updatedUser);
-        return updated != null ? 
-                new ResponseEntity<>(updated, HttpStatus.OK) : 
+        return updated != null ?
+                new ResponseEntity<>(updated, HttpStatus.OK) :
                 new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    // Delete a user
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable int id) {
+    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
         boolean isDeleted = userService.deleteUser(id);
-        return isDeleted ? 
+        return isDeleted ?
                 new ResponseEntity<>("{\"message\": \"User deleted successfully\"}", HttpStatus.OK) :
                 new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    
     @GetMapping("/paginated")
     public ResponseEntity<Page<User>> getPaginatedUsers(
-            @RequestParam(defaultValue = "0") int page, 
+            @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size) {
         Page<User> paginatedUsers = userService.getUsersByPage(size, page);
         return new ResponseEntity<>(paginatedUsers, HttpStatus.OK);
     }
 
-   
     @GetMapping("/paginated/sorted")
     public ResponseEntity<Page<User>> getPaginatedAndSortedUsers(
-            @RequestParam(defaultValue = "0") int page, 
-            @RequestParam(defaultValue = "10") int size, 
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size,
             @RequestParam String field) {
         Page<User> paginatedAndSortedUsers = userService.getUsersByPageAndSort(size, page, field);
         return new ResponseEntity<>(paginatedAndSortedUsers, HttpStatus.OK);
